requirements.txt
alembic==1.5.4
beautifulsoup4==4.9.3
bidict==0.21.2
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
digi-xbee==1.3.0
dnspython==2.1.0
email-validator==1.1.2
eventlet==0.30.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.6.0
Flask-Script==2.0.6
Flask-SocketIO==5.0.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gevent==21.1.2
greenlet==1.0.0
gunicorn==20.0.4
idna==2.10
install==1.3.4
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.4
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
PyJWT==2.0.1
pyserial==3.5
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-engineio==4.0.1
python-socketio==5.1.0
requests==2.25.1
six==1.15.0
soupsieve==2.1
SQLAlchemy==1.3.22
srp==1.0.16
urllib3==1.26.4
websocket-client==0.58.0
Werkzeug==1.0.1
WTForms==2.3.3
zope.event==4.5.0
zope.interface==5.3.0
